Move cmd/fluxctl/install to install/
