+++ install/generate.go
