+++ install/templates/flux-secret.yaml.tmpl
