+++ install/templates/memcache-dep.yaml.tmpl
