+++ install/templates/memcache-svc.yaml.tmpl
